services/user-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
PyMySQL==1.1.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
email-validator==2.2.0
